+++ codex-rs/command/BUILD.bazel
@@ -1,0 +1,7 @@
+load("//:defs.bzl", "codex_rust_crate")
+
+codex_rust_crate(
+    name = "command",
+    crate_name = "codex_command",
+    compile_data = ["src/command_safety/powershell_parser.ps1"],
+)
