+++ codex-rs/command/src/command_safety/is_dangerous_command.rs
