+++ codex-rs/command/src/command_safety/is_safe_command.rs
