+++ codex-rs/command/src/command_safety/mod.rs
