+++ codex-rs/command/src/command_safety/powershell_parser.ps1
