+++ codex-rs/core/src/lib.rs
@@ -10,7 +10,6 @@ pub mod api_bridge;
 mod apply_patch;
 mod apps;
 pub mod auth;
-pub mod bash;
 mod client;
 mod client_common;
 pub mod codex;
@@ -22,7 +21,6 @@ pub use codex_thread::ThreadConfigSnapshot;
 mod agent;
 mod codex_delegate;
 mod command_canonicalization;
-mod command_safety;
 pub mod config;
 pub mod config_loader;
 pub mod connectors;
@@ -52,10 +50,8 @@ mod memories;
 mod mentions;
 mod message_history;
 mod model_provider_info;
-pub mod parse_command;
 pub mod path_utils;
 pub mod personality_migration;
-pub mod powershell;
 mod proposed_plan_parser;
 mod sandbox_tags;
 pub mod sandboxing;
@@ -133,11 +129,14 @@ mod state;
 mod tasks;
 mod user_shell_command;
 pub mod util;
+pub use codex_command::bash;
+pub use codex_command::is_dangerous_command;
+pub use codex_command::is_safe_command;
+pub use codex_command::parse_command;
+pub use codex_command::powershell;
 
 pub use apply_patch::CODEX_APPLY_PATCH_ARG1;
 pub use client::X_CODEX_TURN_METADATA_HEADER;
-pub use command_safety::is_dangerous_command;
-pub use command_safety::is_safe_command;
 pub use exec_policy::ExecPolicyError;
 pub use exec_policy::check_execpolicy_for_warnings;
 pub use exec_policy::load_exec_policy;
